Refactored structure
refactored structure for signin and signup pages
