+++ frontend/src/components/SignIn/SignIn.css
@@ -14,7 +14,7 @@ body{
     display: flex;
     flex-direction: column;
     align-items: center;
-    background-color: #818181;
+    background-color: var(--clr-light-3);
     height: 100vh;
 }
